chore: move TokenCard to component folder
